--- src/components/SolanaBlinkRenderer/utils/constants.ts
@@ -1,1 +1,0 @@
-export const noop = () => {};
