--- src/components/SolanaBlinkRenderer/utils/type-guards.ts
@@ -1,3 +1,0 @@
-export const isSignTransactionError = (
-  data: { signature: string } | { error: string },
-): data is { error: string } => !!(data as any).error;
